Fix optionalness and simplify package layout
