+++ build.sbt
@@ -6,12 +6,13 @@ crossScalaVersions := Seq("2.10.4", scalaVersion.value)
 scalacOptions ++= Seq("-optimize", "-Ywarn-dead-code")
 
 libraryDependencies ++= Seq(
-  "com.socrata" %% "socrata-curator-utils" % "1.0.1" % "optional",
-  "com.socrata" %% "socrata-http-client"   % "3.5.0" % "optional",
-  "com.socrata" %% "socrata-http-jetty"    % "3.5.0" % "optional",
-  "com.socrata" %% "socrata-http-server"   % "3.5.0" % "optional"
-)
+  "com.rojoma"  %% "simple-arm-v2"         % "[2.1.0, 3.0.0)" % "optional",
+  "com.socrata" %% "socrata-curator-utils" % "[1.0.1, 2.0.0)" % "optional",
+  "com.socrata" %% "socrata-http-client"   % "[3.5.0, 4.0.0)" % "optional",
+  "com.socrata" %% "socrata-http-jetty"    % "[3.5.0, 4.0.0)" % "optional",
+  "com.socrata" %% "socrata-http-server"   % "[3.5.0, 4.0.0)" % "optional"
 
+)
 val TestOptionNoTraces = "-oD"
 val TestOptionShortTraces = "-oDS"
 val TestOptionFullTraces = "-oDF"
