+++ src/main/scala/com.socrata.testcommon/UnusedSugarExtras.scala
@@ -1,0 +1,7 @@
+package com.socrata.testcommon
+
+trait UnusedSugarExtras extends UnusedSugarCommon {
+}
+
+object UnusedSugarExtras extends UnusedSugarExtras {
+}
