+++ src/main/scala/com.socrata.testcommon/mocks/StaticHttpClient.scala
@@ -1,11 +1,11 @@
-package com.socrata.test.common.mocks
+package com.socrata.testcommon.mocks
 
 import java.io.Closeable
 
 import com.socrata.http.client._
 import com.socrata.http.server.responses._
 
-class StaticHttpClient(val respBuilder: (SimpleHttpRequest => Response)) extends HttpClient {
+class StaticHttpClient private (val respBuilder: (SimpleHttpRequest => Response)) extends HttpClient {
   val close: Unit = ()
 
   private def rawResp(r: Response) = new RawResponse with Closeable {
