+++ src/main/scala/com.socrata.testcommon/mocks/util/AcknowledgeableInputStream.scala
@@ -1,4 +1,4 @@
-package com.socrata.test.common.mocks
+package com.socrata.testcommon.mocks.util
 
 import java.io.{ByteArrayInputStream, InputStream}
 import java.nio.charset.{Charset, StandardCharsets}
@@ -6,7 +6,7 @@ import java.nio.charset.{Charset, StandardCharsets}
 import com.socrata.http.common.util.Acknowledgeable
 
 /** Will return user provided data to its consumers. */
-class AcknowledgeableInputStream(payload: Array[Byte])
+class AcknowledgeableInputStream private (payload: Array[Byte])
     extends InputStream
     with Acknowledgeable {
   def this(payload: String, charset: Charset) = this(payload.getBytes(charset))
@@ -20,8 +20,12 @@ class AcknowledgeableInputStream(payload: Array[Byte])
 }
 
 object AcknowledgeableInputStream {
-  def apply(payload: Array[Byte]): AcknowledgeableInputStream = new AcknowledgeableInputStream(payload)
-  def apply(payload: String, charset: Charset): AcknowledgeableInputStream = apply(payload.getBytes(charset))
-  def apply(payload: String): AcknowledgeableInputStream = apply(payload, StandardCharsets.UTF_8)
-  def apply(): AcknowledgeableInputStream = apply(Array.empty[Byte])
+  private[mocks] def apply(payload: Array[Byte]): AcknowledgeableInputStream =
+    new AcknowledgeableInputStream(payload)
+  private[mocks] def apply(payload: String, charset: Charset): AcknowledgeableInputStream =
+    apply(payload.getBytes(charset))
+  private[mocks] def apply(payload: String): AcknowledgeableInputStream =
+    apply(payload, StandardCharsets.UTF_8)
+  private[mocks] def apply(): AcknowledgeableInputStream =
+    apply(Array.empty[Byte])
 }
